0.1.5 Submited the game
Game is submitted but not fully polished, because of the lack of time. Game will get an update/s in next days, as well as readme.

Added:
MainMenu with settings and credits.

2 gameovers, PlayerWin and MinotaurWin.

Fader for transitioning between scenes.

Portal for player to escape from the maze.

Audio system with music working smoothly between all scenes.